Ensure space between tag rich links and showcase img
On wider breakpoints, showcase images in the article body
intrude on the left column. Rich links that are automatically
added to the article based on tag id may clash with the
showcase image caption, causing the caption to be hidden. This
looks dreadful and may cause copyright issues.

This fix ensures automatic rich links can't be added 200px
below a showcase image element. This is limited to leftCol,
desktop and wide breakpoints as mobile breakpoints are unaffected.