Strip host from richLinkId
As part of the CAPI response, frontend receives ids of rich links
that ought to be automatically added to an article. These ids
include the host (e.g. https://www.theguardian.com) which causes
problems when running in the application in non-PROD environments.

This change strips the host to make the richLinkId host-agnostic.
This ensures the logic in rich-links.js correctly inserts tag
rich links.