Overlay media image on Twitter with content age
Old articles may be shared on Twitter, which could give readers a
misleading picture of a news story.

This change overlays older articles with a PNG that displays the year
the article was published, giving more context to the reader and
allowing them to disregard stories that are no longer relevant.