Fix microdata for film reviews
Google doesn't recognise article[scope|prop|..] here it seems
(and I can't find any explanation of what they are) so I've
switched to item* and also added some missing fields that Google
requires (image and author).

Longer-term (and in DCR we do this) we should just use JSON+LD as
it is much easier to generate and maintain.