Enable upscaling on opengraph images
Somehow I forgot to enable this when adding the original code.

At the moment we are getting a lot of warnings on structured data
with Google for AMP because we are serving images that are below
their recommended size (1200px). But outside AMP it is worthwhile
too in order to ensure we meet FB and Google recommended sizes for
regular content too.