Fallback when structured data image uses unexpected host
Typically we expect images to be served from the image service. In
particular, this helps us ensure they are large enough to satisfy
Google and other party size requirements.

For unknown cases (rare and usually older content) we simply use
the fallback logo instead.

As part of this change, the alternative ratio images (1x1 etc.)
have been removed as they are don't contribute to validation at
this time.