Add event triggers for videoPlaying and videoStopped states
Whenever a video is playing on a page we need to keep attention time ticking upwards
and this can be done by integrating the new youtube library code into the existing
tracker-js code for videos.
Currently this looks for a DOM event coming from `document.body` and so by triggering
this same event based on if any player on the page is in a `playing` state is potentially
the simplest way to do this without rewriting the way tracker-js monitors videos.