Geolocate users by latlong from Fastly for weather
This is a part of a larger piece of work to move over to a new Fastly
geolocation API, which is required in order to support IPv6.

In order to not leak too detailed information to Accuweather we reduce
the accuracy of the latitude and longitude we get from Fastly to 2
d.p. (according to https://en.wikipedia.org/wiki/Decimal_degrees that
gives town/city accuracy, so should be plenty for the weather).

Note that we're still varying the cache for responses to this endpoint
on country, region and city even though the data will generally be
fetched from accuweather based on lat/long. We don't want to vary based
on lat/long as it will fragment the cache too much.