Handle interactive atoms on fronts
When the embed type is "interactive", conditionally render associated
HTML, CSS and JS. This is pressed into the front and then rendered by
facia if present.

The interactive atom data is fetched using the CAPI client, as
recommended by the editorial tools team. Currently this means extracting
the atom ID from the URL in the fronts json. On CODE I discovered that
when using internal.content.guardianapis.com a double forward slash at
the start of the path breaks things. This happens if the atom ID starts
with a forward slash, so remove any leading forward slash during the ID
extraction. In future the fronts JSON will contain the atom ID directly
so we won't have to do this extraction.