Fix crosswords on IE11 and Safari iOS 9
Crosswords are failing to load on IE11 and Safari on iOS 9. It looks
like this is because we’re serving incorrectly cached polyfills in some
cases, so those browsers don’t get the polyfills they need. Fix this by
incrementing the cache counter to bust the cache (note that this doesn’t
address the root cause of _why_ we’re serving incorrect data from the
cache, which we don’t fully understand yet).