Don’t press front when interactive atom fails
Only recover from failure when a json.html press fails. We want to treat
an interactive atom failure in the same way as an article fetch failure,
since both are from CAPI. This means we won’t end up with an unenriched
card in the front. Also bail out of pressing if the content for the
interactive atom isn't what we expect.