New live blog styling for dynamic containers
Highlights:

* Hide image in dynamic container when live updates are on (desktop
  only).
* Show two updates at a time + circles and connecting line

Notes:

It's a bit of a hack to get the image appearing in liveblogs for
standard card in this layout:

+========+====+
|        |live|
|        |blog|
+========+====+

Generally we want to hide the image for standard cards when we’ve got a
live blog with updates. However we can show the image when it’s in a
column to the right. I’m not sure about using the layout classes here
to determine this. Probably better to add a specific class for this
scenario in Scala. When there is a stack of cards to the right it’s okay
to show the image, because it looks like we don’t show live updates in
this case anyway.

The non-dynamo live update behaviour adds a delay before animating. In
the case of the fade in/out in a dynamo, this leaves an empty section
for a period before the update is faded back in, which looks a bit odd.

We use shorter live update text times in dynamo. Wrap the prefix of the
update time (e.g. "Last update", "Updated") in a span with a class so we
can hide in the dynamo. The casing is a little weird (we’ll now start
with a lower case "just now") but I think that’s okay since we’re
already explicitly lowercasing the "yesterday" case.