Hack to get the image appearing in liveblogs for standard card in this layout:
+========+====+
|        |live|
|        |blog|
+========+====+

Generally we want to hide the image for standard cards when we’ve got a
live blog with updates. However we can show the image when it’s in a
column to the right. I’m not sure about using the layout classes here
to determine this. Probably better to add a specific class for this
scenario in Scala. When there is a stack of cards to the right it’s okay
to show the image, because it looks like we don’t show live updates in
this case anyway.