Use the atomId from fronts to fetch interactive
This replaces the previous workaround which was to extract the ID from
the CAPI URL.

Also includes some changes to the way we handle various error scenarios
for interactives. If for any reason the interactive fails we want to
halt the pressing of the front. I've added some tests to cover/document
these scenarios.