Reduce CAPI client max retry attempt count
We’ve seen a few scenarios lately where CAPI is under strain and
starts returning errors. In these cases frontend will retry making the
request which seems to largely have the effect of kicking CAPI while
it’s down.

In the context of an HTTP request we think that retrying multiple times
probably isn’t what we want. There is a 2 second limit on requests to
frontend. If this limit is reached the request is terminated. Combined
with CAPI’s circuit breaker logic this means that retrying with a delay 
is unlikely to yield a successful result within the 2 second window.

It isn’t possible to return to the previous behaviour of not-retrying at
all, the closest we can get right now is to reduce the retry count limit
to 1 (i.e. an initial request and a single retry if the first request
fails).

Going forward we think we probably _do_ want an exponential backoff with
retries and a limit on the total number of in-flight requests to CAPI
for offline jobs, where response time to an end user isn’t a concern.
This will require splitting the ContentApiClient implementation to
handle both scenarios.