Add call to DiscussionAPI in Identity PublicProfileController (#22212)
We currently show a public profile for all Guardian users, whether or not they have commented. If they have not commented, this profile is blank. This service will allow to check whether or not a user should have a public profile and allow us to server a 404 page if not.

* Add a lightweight service to frontend/identity to make a serverside GET requests to DiscussionApi (DAPI)

* Handle DAPI response including 404 USER_NOT_FOUND - a Guardian User will only be created in Discussion when they comment or recommend a comment for the first time

* Add a fx 'userHasPublicProfile' to call DAPI GET /profile/<id>/stats and check if a user should have a public profile, based on whether or not they have commented.

* Amend config files to use new service in Identity

* Use DiscussionApiService.userHasPublicProfile in Identity PublicProfileController