Set remote render test to 0%
The motivation here is that we need to turn the test switch 'On'
to be able to test on PROD, but we want to limit the test to
opt-in only initially to verify it is working as expected.

Unfortunately, we don't have a good mechanism to opt-in currently
unless the relevant switch is set to 'On'.