Geolocation getSync can return null
The types here aren’t quite right, which caused confusion elsewhere
since we were always expecting to get a country code back from getSync.
The scenario in which it can return null is if there’s no country
code in local storage and we _do_ have a page.edition, but it doesn’t
map to a geolocation (e.g. INT).

This had a ripple effect in a bunch of other places, so I updated types
in a few spots. I think in all cases given our usage of country code
it’s okay if the country code is empty.