Mute youtube videos to autoplay on Android
In order for embedded videos to autoplay on Android devices
we need to mute the video on start. Implementing this here is
intended only for use in specific US paid content pages.

Another way of doing this would be to modify the `<video>` element
to include `autoplay` and `muted` and this would prevent mutation
of the iFrame `src` property. However, this doesn't seem to be easily
achievable but could be the focus of future improvements.

See https://developers.google.com/web/updates/2016/07/autoplay

Co-authored-by: Josh Buckland <buck06191@gmail.com>