Remove old code for handling adslot resize
On slow connections, ignoring iframe resize events causes adverts to
display poorly. The original decision to bypass resizing was due to
issues with the adslot resizing _as well as_ the iframe.

Here we have removed the adslot resize event and the bypassing of the
overall resize event for specific ad slot types.

Co-authored-by: Ioanna Kyprianou <ioanna.kyprianou.contractor@guardian.co.uk>