Delay geo fetch for remote epic (#22322)
The AB test code is called *before* the geolocation is set.
Depending on which one executes first (they are both async) the
geolocation may be more or less correct.

Note, geo defaults to an edition-based logic if the local
storage has not been set yet.

Fixing this would require substantial code reorganisation. By
reloading the value as late as possible, we give ourselves the
best chance of getting the right geo.