Implement youtube testing
Implement initial youtube tests for iframe behaviour and
muting of iframes.

This includes a reformatting of the `getIFrameBehaviourConfig` method
to take a `HTMLIFrameElement` as an argument rather than a string.

This makes the code easier to test as it doesn't rely on passing
through the global `players` variable from `youtube.js`.