Fix live updates background colour issue
This was only affecting the two card dynamo layout.

There was also a specificity issue which I've now fixed which ensures we
get the hover state background colour of the live blocks correct
(hovering on a sublink was causing the background of live blocks to be
slightly different than the border-bottom and ellipses background).