Support interactive atoms on Amp
https://trello.com/c/gNgwpVUb/881-render-interactives-on-their-unique-end-points-to-be-used-in-amp-iframe-and-replace-srcdoc-with-src-for-amp-iframe

Interactive Atoms on Amp should be rendered in an iframe with a src
attribute, instead of srcdoc as they are right now. This PR adds support
for interactive atoms to the /embed/atom endpoint. An InteractiveAtom
now maps to InteractiveUrlBlockElement for Amp. The url is an instance
of the /embed/atom endpoint.