Commercial Lifecycle now uses its own Execution Context
We've been seeing performance problems on the commercial app in
Frontend at start time. The webapp's latency spikes significantly,
which has a knock-on effect on the router, which in turn affects all
the apps in frontend.

This change introduces a bespoke execution context for the Commercial
Lifecycle class. This means the lifecycle work is done in a separate
EC, isolating it from the EC used to serve requests. The choice of EC
(FixedThreadpool) also serves to throttle the execution of these jobs,
which should stop them consuming so many resources that the webapp's
ability to serve timely requests is diminished.