Introduce ability to ignore can run for a forced test
Specifically, if you want to force a client-side AB test via the
URL, you can now also force it to always run by ignoring the
canRun.

The initial motivation is to allow easily checking rendering of
different epic variants, but it is expected to be useful in other
cases too.