Introduce remote epic (all) variants test
The aim here is to serve the epic from the remote service for
a subset of users (determined by the canRun) and see how variants
perform compared to those served locally.

Success would be that the remote variants perform the same as the
local ones.

The remote service includes an `isRemote` parameter in the cta
link that we can use to compare results.

The pageview tracking event also contains the remote epic test
metadata if that branch runs, which can be used to filter insert
and view events (these join onto this table in the Data Lake).

Note, we will have to update the contributions Tableau dashboards
to account for the above.

As part of this PR, some refactoring has been done to move out
some of the event code. As part of this, the 'success' and
'impression' properties on EpicAbTest have gone - this simplifies
things as we no longer need events. In addition, it wasn't
possible to use this interface for the remote epic, as the required
metadata isn't available at construction but would need to be
passed in (from the Contributions Service API result), which the
interface does not support.