Fix unintentional double back event from lightbox
Opening, closing and then re-opening the lightbox was causing keyboard
events to be re-bound without correctly unbinding the previously bound
handler. This meant that when exiting the lightbox for the second time
using the escape key, the event was being handled twice, resulting in
the browser going back twice, instead of just once to the article.

I think this was because the bean.off and on keydown handling was using
anonymous arrow functions, which will be unique every time. Instead we
now bind this to handleKeyEvents in the constructor and can use
this.handleKeyEvents directly, no arrow function required.