Fix problem with slotReady being undefined
Errors were occurring for the `slotReady` promise when the
`iasAdTargeting` switch was turned off. This was found to be due to the
`slotReadyz variable being initialised to `undefined` and then never set
as a promise outside of the `iasAdTargeting` closure. This would then
prevent adverts from being shown on screen.

This change intialises `slotReady` as a `Promise` instead of as an
`undefined` variable.

Tested against localhost with the `iasAdTargeting` switch both on and
off.