Add AtomEmbedBlockElement removing seperate markup and url
We don't need to handle atomEmbedMarkupBlockElement and atomEmbedUrlBlockElement different in DCR. In fact, we can just use the URL in AMP and the html, css and js to srcdoc in dotcom.

We also create an InteractiveBlockElement proper, as previously it was just writing it to the AtomEmbedUrlBlockElement, and we wanted to do it properly.