Modify data-link-name based on state
We currently don't differentiate bthe `data-link-name` based on whether
a user is signed in or if it is the first/second stage of the non-signed
in sign up process.

This introduces a set of functions to modify the `data-link-name` based
on whether it is the first/second stage of newsletter sign up or if the
user is already signed in.