Increase newsletter tracking granularity
In order to properly track the newsletter sign ups at a more
granular level, we need to use the `identityName` for the
`data-link-name`. To do this and to maintain current functionality
I've added a new optional field to the checkboxes used such that
a newsletter will use the `identityName` whilst consents and sms
will use the existing `name` field.