Add tag links to DCR (take two)
Note, the existing code in HtmlCleaners was repurposed but not
directly as the Elements model requires a new approach. This was
also the problem with the previous fix - it ignored the fact that
applying tag links to a paragraph is not stateless but a function
of tag links on previous paragraphs; we only want to link the first
occurrence of a tag.