Prep newsletter embeds for new designs
New designs are being introduced for newsletter embeds and a good amount
of health work is required in order to both maintain the general health
of the codebase and to allow integration of new designs with minimal
reliance on Editorial changes.

Here I've introduced a number of important changes. Firstly, the code
handling the rendering of the footer embeds is now separate from the
code for in article embeds. This should mean that it's easier to
implement specific design changes for each embed type without having
increasing levels of business logic tied into the rendering code.

Secondly, the Identity model has been updated to include an `EmailEmbed`
field which contains the copy that will be displayed for each in-article
embed. As a result, there is no need to pass this information to the
iframe as attributes. In addition, there is an endpoint per newsletter
for the success state. This allows specific copy for any one newsletter
following submission.

There remain a number of changes that need to be made following this
specific commit:

 - Cleaning up of the codebase to remove unused healthchecks etc., that
   rely on the previous format and endpoints. This will then allow us to
   remove any of the `V2` methods in the controller. These only exist at
   present to allow the full project to compile.
 - Styling changes. There are still shared styles between the footer and
   in-article embeds. This is causing a number of issues, especially on
   the submission success endpoint. In particular, we need black text
   in articles and white text in the footer.