Changes following code review
Make some changes following code review including some basic
refactoring.

Biggest change is reworking the submission endpoint and method to
extract the `listName` from the form directly instead of passing it as a
URL path component. This then necessitated changing the result endpoints
to remove the reliance of non-success states on receiving a `listName`
from the form; an error state may not have a form to extract this from,
and the `listName` was never used in the returned fragment for those
states.