Explicitly define Json deserialisation
Because the Newsletter API only reveals a subset of the fields on the
`EmailNewsletter` case class we needed to explicitly define the
case class for that type and add case classes to use that in place
EmailNewsletter for all the further downstream case classes that also
use it i.e. `GroupedNewsletter`.