Add carousel Images to onwards content
Adds a new field to the onwards content to pass through URLs to specific
image sizes for the different breakpoints we expect to use in the
Carousel.

I've not modified the existing image field as I don't want to break
backwards compatibility, but I think that there is scope to change this
to create a set of images for different breakpoints even outside of the
Carousel.