Use the IDAPI for the All Newsletters page
Changes to the newsletters data model have so far required that we
release a new version of Identity and then bump the version in Frontend.
This has had the main side effect of preventing large changes around
things like Play versions in Identity as they are incompatible with
Frontend.

By here moving to using the Newsletters API that was previously
introduced for the Email Sign Up Embeds, we allow Frontend to move away
from requiring Identity version bumps every time a Newsletter is
updated. Instead the information is made available via the API.

I've followed the same principles as done previosuly by creating an
Agent and Lifecycle method to cache and refresh this data. I've then
updated the template to use this data as needed.