Remove DCR experiment and switch (#23661)
DCR is close to 100% now and is certainly here to stay.

Note, ?dcr=true can still be used to force DCR on for some
unsupported article types. But it is no longer possible to
opt-out of DCR for supported pages.

As part of this change, several article-related Frontend tests
have been removed. These are better performed by DCR itself
now that we are close to 100% of vanilla articles migrated:

* avoids coupling our build process to DCR availability
* avoids duplication of tests between DCR+Frontend
* takes us closer to the slimline Frontend we want