Implement Snyk Github Actions
We want/need Snyk on our repos and projects, but there are a number of
drawbacks to using the existing Github integration, namely that it can
block deployments due to vulnerabilities unrelated to the code changes
in place.

The alternative introduced here is to use the Github Action instead.
This doesn't block anything, the config is version controlled and it
still gives devs the relevant info.