Temporarily force documentaries into DCR
Most documentaries are already switched over, but some have the
opt-out tag. Unfortunately, removing this tag is pretty difficult
at the moment. Composer are changing things to make this easier but
in the meantime we can ignore the tag for documentaries.

Note, this is important as documentaries share the same rendering
code so the DCR-specific fixes that are live are resulting in visual
issues (albeit minor) on the opted out ones that are still served by
Frontend.