feat: remove guest password from the identity onboarding flow
As part of the work done in the Identity migration project, we want new
identity accounts to have strong passwords, as well as to go through
the new onboarding flow on [gateway](https://github.com/guardian/gateway)

However since this flow in this project is still in use by email
newsletters, and the migration isn't trivial. We've decided that it's
safer to remove the password form and setting ability from this flow
while leaving the rest of it in place, and we'll migrate this at a later
date.