serve pressed articles based on pressed interactive data
At the time of committing it is not possible to batch tag pressed interactives
or programmatically add a tag. While this solution is being worked on, we will
keep a list of pressed interactives that Visuals have agreed to show to readers.
When we have the ability to add tags to pressed interactives our solution will
change. For now we also retain the experiment in the controller to enable us to
quickly hide pressed content in the event of unexpected errors.