add interactive pressing to admin tool (#24173)
We add a page to the frontend admin tool to allow admins to press an interactive.
The new page takes a URL and both presses & cleans the interactive. Note that, at
the time of committing, there is not yet a programmatic solution for tagging pressed
articles. As a result, there is a subsequent manual step to display pressed content
to readers.