update interactive picker to determine if a pressed article should be served
Moving the logic to determine whether a pressed article should be served
into the interactive picker. For the intermediate solution (retaining a list
or pressed paths in config) this means we make an extra call to CAPI, but in
the long term we'll need to do this anyway (so we can check tags). Moving
the logic into the interactive picker probably makes the controller code
more elegant too.