Fix bug in spacefinder nearby candidate filtering
Fixes a bug in the "nearby candidate filtering" mechanism in spacefinder.

This bug was caused by a variable (`previousAllowedCandidate`) being defined in the global scope and not being managed correctly. As a result, state would bleed between consecutive calls to `filterNearbyCandidates`. This sometimes led to strange behaviour.

In particular, it sometimes meant candidate paragraphs which should have been marked as winning paragraphs were instead excluded on the basis they were too close to a previous winning paragraph.

The fix is to stop `filterNearbyCandidates` relying on global state to keep track of the last winning paragraph. Instead, this responsibility was moved into the `filter` function in spacefinder, which now passes the last winning paragraph into `filterElement` as an optional second argument.