Don't wait for lazily-loaded images to load
Don't wait for lazily-loaded images to load before running spacefinder.

Since September 2020, all images in articles (except the main one) are configured to load lazily by default. This means they are only fetched when the user scrolls sufficiently close to them. Unfortunately, spacefinder relies on all images being loaded before executing.

In practice, this means that for any article containing at least one image (excluding the main one), spacefinder waits for 5 seconds before running, which is the loading timeout, unless the user happens to scroll to all images before then.

On desktop, the situation is worse because the rules are applied twice, firstly to insert an `inline1` ad, and secondly to insert `inline2`, `inline3` etc. ads, as appropriate. This means execution is stalled by 10 seconds in the worst case.

This is particularly apparent on long reads where the user is unlikely to scroll to all images within 10 seconds.

Given lazy-loading is now the default behaviour for images on articles, there is a case to be made for removing the waiting mechanism from spacefinder altogether. However, we would need to consider other uses of spacefinder (e.g. liveblog adverts) before making that change, so I am leaving it for now.

See: https://github.com/guardian/dotcom-rendering/pull/1882