Ignore rich links when placing `inline1` ads
Ignore rich links when placing `inline1` ads on desktop.

Currently `inline1` ad placement is subject to strict rules which mean any HTML element which isn't one of the following:

- `p`
- `h2`
- `.ad-slot`

blocks the placement of `inline1` ads. This doesn't make sense for rich links on desktop, which appear left of the article.

The fix is to add rich links to the allow list. Unfortunately, we don't currently have a good way to identify rich links so this PR relies on `data-spacefinder-component` being set correctly in a separate PR.