Fix spacefinder in liveblogs (#24779)
* Re-run spacefinder when DCR liveblogs are updated

DCR emits a `liveblog:blocks-updated` event whenever new blocks are added to liveblogs. We need to listen for this event and rerun spacefinder accordingly.

* Fix spacefinder bug in liveblogs

Until recently `insertAds` always returned `undefined`. As a result, `startListening()` never got executed, which meant spacefinder never re-ran after new liveblog blocks were added. This was addressed in this PR:

https://github.com/guardian/frontend/pull/24599

At the time, we preserved this (broken) behaviour by manually setting `result` to `undefined`.

This commit introduces an experiment `SpacefinderOkr4HandleNewLiveblogBlocks` which fixes the bug for users in the `variant` group. This will allow us to measure any revenue uplift associated with this change.

* rebuild commercial bundle graph

* remove event listener in stopListening

* Revert "Fix spacefinder bug in liveblogs"

This reverts commit 3a9026e534fca0dc62689f2d91fa1bb929961f99.

* fix bug preventing spacefinder from rerunning

* Update standalone.commercial.ts.json

* break ad refresh for users in variant of spacefinderOkrMegaTest

* rebuild commercial bundle graph

* mock ophan in liveblog-adverts.spec.js

* remove `eslint-disable-next-line`s

* only. the control group of experiment should have bug fix disabled

* update comment

Co-authored-by: Simon Byford <simon.byford@guardian.co.uk>

* Ignore result when determining whether or not to listen for more blocks

* injectNewBlocks dispatches liveblog:blocks-updated

* Remove mediator listener from liveblog adverts

Co-authored-by: Zeke Hunter-Green <zekehuntergreen@gmail.com>