Add container to `inline2+` ad slots (again)
Second attempt at adding a container div to `inline2+` ad slots. It follows on from a previous attempt:

https://github.com/guardian/frontend/pull/24917

Unfortunately this caused a GPT issue which broke inline2 ads and was quickly disabled in this PR:

https://github.com/guardian/frontend/pull/24930

The issue occurred because we passed the container div into `loadAdvert` instead of the ad element itself.

The fix is to move the responsibility of adding a container div from `createAdSlot` into `insertAdAtPara`