Fix `onInteractivesLoaded` race condition
Currently when checking to see if all interactives have loaded before running spacefinder, we wait for their associated iframe to have `readyState === 'complete'`.

I propose we instead listen for a 'set-height' message to be sent from the iframe. This happens very soon after the iframe is loaded but not immediately. This leads to a race condition whereby spacefinder sometimes is run before the iframe has resized and therefore with incorrect page measurements.