Delete superfluous `implicits.Collections` class
In the course of upgrading the Frontend codebase to Scala 2.13 we came
across the `implicits.Collections` class, and had trouble updating it
for Scala 2.13: https://github.com/guardian/frontend/pull/25190#discussion_r913972996

It turns out that we can _delete_ this class with the Scala 2.13
upgrade, rather than have to update it, which is nice!

To reduce the size of the Scala 2.13 upgrade PR, we first removed
every use of `implicits.Collections` that we could while still under
Scala 2.12: https://github.com/guardian/frontend/pull/25226

This commit gets rid of `implicits.Collections` altogether. At this
point, all `implicits.Collections` has in it is the code to add a
`distinctBy()` method onto standard Scala collections - this was
introduced November 2012 with https://github.com/guardian/frontend/pull/263,
but Scala 2.13 has a new built-in method that works the exact same way!
So we can delete the Frontend implementation.

https://www.scala-lang.org/api/2.13.x/scala/collection/Seq.html#distinctBy%5BB%5D(f:A=%3EB):C