Add (failing) test case for intra-collection dedup
Adds a test to check that backfill trails are deduped against curated
trails from the same collection, as well as against previous collections on the page.

This test case fails due to a bug in the current code, in anticipation
of a fix which should be implemented in the next commit.