Scala 2.13: Refactor cricket match summary template
This commit removes duplication in the cricket match summary template,
which is embedded into the /sport/cricket/match/.../...-team.json endpoint,
delivering a dynamically loaded HTML chunk summarising the scores on to
Cricket articles like this:

https://www.theguardian.com/sport/live/2022/jul/31/england-v-south-africa-third-mens-t20-live-cricket

(the corresponding json endpoint for that article is
https://api.nextgen.guardianapps.co.uk/sport/cricket/match/2022-07-31/england-cricket-team.json
...you have to work out the date of the article and then transplant it into the url for the
json endpoint)

We're removing the duplication because in general DRY
(https://en.wikipedia.org/wiki/Don%27t_repeat_yourself) is a good thing,
and the intricacies of cricket need as much clarity as possible for
non-cricket-following developers! The reason we're looking at this
code though (and trying to understand all the cricket-related-logic!)
is because of a compilation error when compiling under Scala 2.13.

The Scala 2.13 compiler is stricter about non-exhaustive matches in
case-pattern-match expression than Scala 2.12 was, and this is actually
quite a good thing - note that there was recently a bugfix in this bit
of code because an unanticipated case in the PA data caused HTTP 500
errors: https://github.com/guardian/frontend/pull/24499

In the case of the `@score()` reusable template block, the Scala 2.13
compiler couldn't verify that all cases were covered - even though
`!innings.closed` (meaning the innings are ongoing) is the logical
complement of (`innings.declared` + `innings.forfeited` + `innings.allOut`),
the compiler couldn't verify that without a knowledge of cricket!
It's simpler just to catch all the specific 'closed' cases we care
about, and then catch the remaining 'open' case with `case _ =>`
to very obviously catch everything that isn't closed.

Co-authored-by: Roberto Tyley <roberto.tyley@guardian.co.uk>